refactor(android): split ReguertaRoot into modular route files
