+++ android/Reguerta/app/src/main/java/com/reguerta/user/presentation/access/ReguertaRoot.kt
@@ -1098,7 +1098,7 @@ private fun RecoverPasswordCard(
     }
 }
 
-private enum class HomeDestination {
+enum class HomeDestination {
     DASHBOARD,
     MY_ORDER,
     MY_ORDERS,
@@ -1551,2090 +1551,6 @@ private fun HomeRoute(
 }
 
 @Composable
-private fun HomeShellTopBar(
-    title: String,
-    canNavigateBack: Boolean,
-    onBack: () -> Unit,
-    onOpenMenu: () -> Unit,
-    onOpenNotifications: () -> Unit,
-) {
-    var isImpersonationExpanded by rememberSaveable { mutableStateOf(false) }
-    Card {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 12.dp, vertical = 10.dp),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            IconButton(onClick = if (canNavigateBack) onBack else onOpenMenu) {
-                Icon(
-                    imageVector = if (canNavigateBack) Icons.AutoMirrored.Filled.ArrowBack else Icons.Filled.Menu,
-                    contentDescription = if (canNavigateBack) {
-                        stringResource(R.string.common_action_back)
-                    } else {
-                        stringResource(R.string.home_shell_menu)
-                    },
-                )
-            }
-
-            Text(
-                text = title,
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.SemiBold,
-            )
-
-            IconButton(onClick = onOpenNotifications) {
-                Icon(
-                    imageVector = Icons.Filled.Notifications,
-                    contentDescription = stringResource(R.string.home_shell_notifications),
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun HomeDrawerContent(
-    member: Member?,
-    currentDestination: HomeDestination,
-    installedVersion: String,
-    onNavigate: (HomeDestination) -> Unit,
-    onCloseDrawer: () -> Unit,
-    onSignOut: () -> Unit,
-) {
-    val drawerScrollState = rememberScrollState()
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(horizontal = 16.dp, vertical = 20.dp),
-        verticalArrangement = Arrangement.spacedBy(16.dp),
-    ) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.Start,
-        ) {
-            IconButton(onClick = onCloseDrawer) {
-                Icon(
-                    imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                    contentDescription = stringResource(R.string.common_action_back),
-                )
-            }
-        }
-
-        Column(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.spacedBy(12.dp),
-        ) {
-            Box(
-                modifier = Modifier
-                    .size(76.dp)
-                    .clip(CircleShape)
-                    .background(MaterialTheme.colorScheme.primary.copy(alpha = 0.14f)),
-                contentAlignment = Alignment.Center,
-            ) {
-                Icon(
-                    imageVector = Icons.Filled.AccountCircle,
-                    contentDescription = stringResource(R.string.home_shell_profile_placeholder),
-                    tint = MaterialTheme.colorScheme.primary,
-                    modifier = Modifier.size(44.dp),
-                )
-            }
-            if (member != null) {
-                Text(
-                    text = member.displayName,
-                    style = MaterialTheme.typography.titleMedium,
-                    fontWeight = FontWeight.SemiBold,
-                )
-                Text(
-                    text = member.normalizedEmail,
-                    style = MaterialTheme.typography.bodySmall,
-                )
-            }
-        }
-
-        Column(
-            modifier = Modifier
-                .weight(1f, fill = true)
-                .verticalScroll(drawerScrollState),
-            verticalArrangement = Arrangement.spacedBy(16.dp),
-        ) {
-            HomeDrawerSection(title = stringResource(R.string.home_shell_section_common))
-            HomeDrawerItem(
-                icon = Icons.Filled.Home,
-                label = stringResource(R.string.home_title),
-                selected = currentDestination == HomeDestination.DASHBOARD,
-                onClick = { onNavigate(HomeDestination.DASHBOARD) },
-            )
-            HomeDrawerItem(
-                icon = Icons.Filled.ShoppingCart,
-                label = stringResource(R.string.module_my_order),
-                selected = currentDestination == HomeDestination.MY_ORDER,
-                onClick = { onNavigate(HomeDestination.MY_ORDER) },
-            )
-            HomeDrawerItem(
-                icon = Icons.AutoMirrored.Filled.Article,
-                label = stringResource(R.string.module_my_orders),
-                selected = currentDestination == HomeDestination.MY_ORDERS,
-                onClick = { onNavigate(HomeDestination.MY_ORDERS) },
-            )
-            HomeDrawerItem(
-                icon = Icons.Filled.CalendarToday,
-                label = stringResource(R.string.module_shifts),
-                selected = currentDestination == HomeDestination.SHIFTS,
-                onClick = { onNavigate(HomeDestination.SHIFTS) },
-            )
-            HomeDrawerItem(
-                icon = Icons.AutoMirrored.Filled.Article,
-                label = stringResource(R.string.home_shell_news_title),
-                selected = currentDestination == HomeDestination.NEWS,
-                onClick = { onNavigate(HomeDestination.NEWS) },
-            )
-            HomeDrawerItem(
-                icon = Icons.Filled.Notifications,
-                label = stringResource(R.string.home_shell_notifications),
-                selected = currentDestination == HomeDestination.NOTIFICATIONS,
-                onClick = { onNavigate(HomeDestination.NOTIFICATIONS) },
-            )
-            HomeDrawerItem(
-                icon = Icons.Filled.Group,
-                label = stringResource(R.string.home_shell_action_profile),
-                selected = currentDestination == HomeDestination.PROFILE,
-                onClick = { onNavigate(HomeDestination.PROFILE) },
-            )
-            HomeDrawerItem(
-                icon = Icons.Filled.Settings,
-                label = stringResource(R.string.home_shell_action_settings),
-                selected = currentDestination == HomeDestination.SETTINGS,
-                onClick = { onNavigate(HomeDestination.SETTINGS) },
-            )
-
-            if (member?.canManageProductCatalog == true || member?.isProducer == true) {
-                HomeDrawerSection(title = stringResource(R.string.home_shell_section_producer))
-            }
-            if (member?.canManageProductCatalog == true) {
-                HomeDrawerItem(
-                    icon = Icons.Filled.Storefront,
-                    label = stringResource(R.string.home_shell_action_products),
-                    selected = currentDestination == HomeDestination.PRODUCTS,
-                    onClick = { onNavigate(HomeDestination.PRODUCTS) },
-                )
-            }
-            if (member?.isProducer == true) {
-                HomeDrawerItem(
-                    icon = Icons.Filled.Inbox,
-                    label = stringResource(R.string.home_shell_action_received_orders),
-                    selected = currentDestination == HomeDestination.RECEIVED_ORDERS,
-                    onClick = { onNavigate(HomeDestination.RECEIVED_ORDERS) },
-                )
-            }
-
-            if (member?.isAdmin == true) {
-                HomeDrawerSection(title = stringResource(R.string.home_shell_section_admin))
-                HomeDrawerItem(
-                    icon = Icons.Filled.Group,
-                    label = stringResource(R.string.home_shell_action_users),
-                    selected = currentDestination == HomeDestination.USERS,
-                    onClick = { onNavigate(HomeDestination.USERS) },
-                )
-                HomeDrawerItem(
-                    icon = Icons.Filled.Add,
-                    label = stringResource(R.string.home_shell_action_publish_news),
-                    selected = currentDestination == HomeDestination.PUBLISH_NEWS,
-                    onClick = { onNavigate(HomeDestination.PUBLISH_NEWS) },
-                )
-                HomeDrawerItem(
-                    icon = Icons.Filled.Campaign,
-                    label = stringResource(R.string.home_shell_action_admin_broadcast),
-                    selected = currentDestination == HomeDestination.ADMIN_BROADCAST,
-                    onClick = { onNavigate(HomeDestination.ADMIN_BROADCAST) },
-                )
-            }
-        }
-
-        HorizontalDivider()
-        TextButton(
-            onClick = onSignOut,
-            modifier = Modifier.fillMaxWidth(),
-        ) {
-            Text(stringResource(R.string.access_action_sign_out))
-        }
-        Text(
-            text = stringResource(R.string.home_shell_version_format, installedVersion.ifBlank { "0.0.0" }),
-            style = MaterialTheme.typography.bodySmall,
-            modifier = Modifier.fillMaxWidth(),
-        )
-    }
-}
-
-@Composable
-private fun HomeDrawerSection(
-    title: String,
-) {
-    Text(
-        text = title,
-        style = MaterialTheme.typography.labelLarge,
-        color = MaterialTheme.colorScheme.primary,
-        fontWeight = FontWeight.SemiBold,
-    )
-}
-
-@Composable
-private fun HomeDrawerItem(
-    icon: ImageVector,
-    label: String,
-    selected: Boolean = false,
-    onClick: () -> Unit,
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clip(MaterialTheme.shapes.medium)
-            .background(
-                if (selected) MaterialTheme.colorScheme.primary.copy(alpha = 0.10f) else MaterialTheme.colorScheme.surface,
-            )
-            .clickable(onClick = onClick)
-            .padding(horizontal = 10.dp, vertical = 10.dp),
-        horizontalArrangement = Arrangement.spacedBy(12.dp),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Icon(
-            imageVector = icon,
-            contentDescription = null,
-            tint = MaterialTheme.colorScheme.primary,
-        )
-        Text(
-            text = label,
-            style = MaterialTheme.typography.bodyLarge,
-            modifier = Modifier.weight(1f),
-        )
-    }
-}
-
-@Composable
-private fun NextShiftsCard(
-    nextDeliveryShift: ShiftAssignment?,
-    nextMarketShift: ShiftAssignment?,
-    deliveryCalendarOverrides: List<DeliveryCalendarOverride>,
-    isLoading: Boolean,
-    members: List<Member>,
-    onViewAll: () -> Unit,
-) {
-    var isImpersonationExpanded by rememberSaveable { mutableStateOf(false) }
-    Card {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp),
-        ) {
-            Text(
-                text = stringResource(R.string.shifts_next_title),
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.SemiBold,
-            )
-            Text(
-                text = stringResource(R.string.shifts_next_subtitle),
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-            if (isLoading) {
-                Text(
-                    text = stringResource(R.string.shifts_loading),
-                    style = MaterialTheme.typography.bodyMedium,
-                )
-            } else {
-                ShiftSummaryRow(
-                    label = stringResource(R.string.shifts_next_delivery),
-                    shift = nextDeliveryShift,
-                    overrides = deliveryCalendarOverrides,
-                    members = members,
-                )
-                ShiftSummaryRow(
-                    label = stringResource(R.string.shifts_next_market),
-                    shift = nextMarketShift,
-                    overrides = deliveryCalendarOverrides,
-                    members = members,
-                )
-            }
-            Button(onClick = onViewAll) {
-                Text(text = stringResource(R.string.shifts_view_all))
-            }
-        }
-    }
-}
-
-@Composable
-private fun ShiftSummaryRow(
-    label: String,
-    shift: ShiftAssignment?,
-    overrides: List<DeliveryCalendarOverride>,
-    members: List<Member>,
-) {
-    Column(
-        modifier = Modifier.fillMaxWidth(),
-        verticalArrangement = Arrangement.spacedBy(4.dp),
-    ) {
-        Text(
-            text = label,
-            style = MaterialTheme.typography.labelLarge,
-            color = MaterialTheme.colorScheme.primary,
-        )
-        Text(
-            text = shift?.toSummaryLine(members, overrides).orEmpty().ifBlank {
-                stringResource(R.string.shifts_next_pending)
-            },
-            style = MaterialTheme.typography.bodyMedium,
-        )
-    }
-}
-
-@Composable
-private fun ShiftsRoute(
-    shifts: List<ShiftAssignment>,
-    shiftSwapRequests: List<ShiftSwapRequest>,
-    dismissedShiftSwapRequestIds: Set<String>,
-    nextDeliveryShift: ShiftAssignment?,
-    nextMarketShift: ShiftAssignment?,
-    deliveryCalendarOverrides: List<DeliveryCalendarOverride>,
-    currentMember: Member?,
-    members: List<Member>,
-    isLoading: Boolean,
-    isUpdatingShiftSwapRequest: Boolean,
-    onRefresh: () -> Unit,
-    onRequestShiftSwap: (String) -> Unit,
-    onAcceptShiftSwapRequest: (String, String) -> Unit,
-    onRejectShiftSwapRequest: (String, String) -> Unit,
-    onCancelShiftSwapRequest: (String) -> Unit,
-    onConfirmShiftSwapRequest: (String, String) -> Unit,
-    onDismissShiftSwapActivity: (String) -> Unit,
-) {
-    var selectedSegment by rememberSaveable { mutableStateOf(ShiftBoardSegment.DELIVERY) }
-    val deliveryShifts = remember(shifts) {
-        shifts.filter { it.type == ShiftType.DELIVERY }.sortedBy { it.dateMillis }
-    }
-    val marketShifts = remember(shifts) {
-        shifts.filter { it.type == ShiftType.MARKET }.sortedBy { it.dateMillis }
-    }
-
-    Column(
-        verticalArrangement = Arrangement.spacedBy(16.dp),
-    ) {
-        Card {
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp),
-                verticalArrangement = Arrangement.spacedBy(8.dp),
-            ) {
-                Text(
-                    text = stringResource(R.string.shifts_title),
-                    style = MaterialTheme.typography.titleMedium,
-                    fontWeight = FontWeight.SemiBold,
-                )
-                Text(
-                    text = stringResource(R.string.shifts_list_subtitle),
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-                Button(onClick = onRefresh) {
-                    Text(text = stringResource(R.string.shifts_refresh_action))
-                }
-            }
-        }
-
-        NextShiftsCard(
-            nextDeliveryShift = nextDeliveryShift,
-            nextMarketShift = nextMarketShift,
-            deliveryCalendarOverrides = deliveryCalendarOverrides,
-            isLoading = isLoading,
-            members = members,
-            onViewAll = onRefresh,
-        )
-
-        ShiftSwapRequestsCard(
-            requests = shiftSwapRequests,
-            dismissedRequestIds = dismissedShiftSwapRequestIds,
-            shifts = shifts,
-            deliveryCalendarOverrides = deliveryCalendarOverrides,
-            members = members,
-            currentMemberId = currentMember?.id,
-            selectedSegment = selectedSegment,
-            isUpdating = isUpdatingShiftSwapRequest,
-            onAccept = onAcceptShiftSwapRequest,
-            onReject = onRejectShiftSwapRequest,
-            onCancel = onCancelShiftSwapRequest,
-            onConfirm = onConfirmShiftSwapRequest,
-            onDismissRequest = onDismissShiftSwapActivity,
-        )
-
-        if (isLoading) {
-            Card {
-                Text(
-                    text = stringResource(R.string.shifts_loading),
-                    modifier = Modifier.padding(16.dp),
-                )
-            }
-        } else if (shifts.isEmpty()) {
-            Card {
-                Text(
-                    text = stringResource(R.string.shifts_empty_state),
-                    modifier = Modifier.padding(16.dp),
-                )
-            }
-        } else {
-            ShiftBoardSegmentSelector(
-                selectedSegment = selectedSegment,
-                onSegmentSelected = { selectedSegment = it },
-            )
-
-            val boardShifts = when (selectedSegment) {
-                ShiftBoardSegment.DELIVERY -> deliveryShifts
-                ShiftBoardSegment.MARKET -> marketShifts
-            }
-
-            if (boardShifts.isEmpty()) {
-                Card {
-                    Text(
-                        text = stringResource(R.string.shifts_empty_state),
-                        modifier = Modifier.padding(16.dp),
-                    )
-                }
-            } else {
-                boardShifts.forEach { shift ->
-                    ShiftBoardCard(
-                        shift = shift,
-                        deliveryCalendarOverrides = deliveryCalendarOverrides,
-                        members = members,
-                        currentMemberId = currentMember?.id,
-                        onRequestShiftSwap = onRequestShiftSwap,
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun ShiftBoardSegmentSelector(
-    selectedSegment: ShiftBoardSegment,
-    onSegmentSelected: (ShiftBoardSegment) -> Unit,
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clip(RoundedCornerShape(24.dp))
-            .background(MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.4f))
-            .padding(4.dp),
-        horizontalArrangement = Arrangement.spacedBy(4.dp),
-    ) {
-        ShiftBoardSegment.entries.forEach { segment ->
-            val isSelected = selectedSegment == segment
-            TextButton(
-                onClick = { onSegmentSelected(segment) },
-                modifier = Modifier
-                    .weight(1f)
-                    .clip(RoundedCornerShape(20.dp))
-                    .background(
-                        if (isSelected) {
-                            MaterialTheme.colorScheme.primary.copy(alpha = 0.16f)
-                        } else {
-                            MaterialTheme.colorScheme.surface.copy(alpha = 0f)
-                        }
-                    ),
-            ) {
-                Text(
-                    text = stringResource(segment.labelRes),
-                    color = if (isSelected) {
-                        MaterialTheme.colorScheme.primary
-                    } else {
-                        MaterialTheme.colorScheme.onSurfaceVariant
-                    },
-                    fontWeight = if (isSelected) FontWeight.SemiBold else FontWeight.Medium,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun ShiftBoardCard(
-    shift: ShiftAssignment,
-    deliveryCalendarOverrides: List<DeliveryCalendarOverride>,
-    members: List<Member>,
-    currentMemberId: String?,
-    onRequestShiftSwap: (String) -> Unit,
-) {
-    val primaryNames = shift.primaryBoardNames(members)
-    val leftLines = shift.leftBoardLines(deliveryCalendarOverrides)
-    val leftAlignment = if (shift.type == ShiftType.MARKET) Alignment.CenterHorizontally else Alignment.Start
-    val canRequestShiftSwap = remember(shift, currentMemberId, deliveryCalendarOverrides) {
-        currentMemberId != null && shift.canBeRequestedBy(currentMemberId, deliveryCalendarOverrides)
-    }
-    val highlightedIndex = remember(shift, currentMemberId) {
-        currentMemberId?.let { shift.highlightedBoardNameIndex(it) }
-    }
-    Card {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp),
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.spacedBy(16.dp),
-            ) {
-                Column(
-                    modifier = Modifier.weight(0.38f),
-                    horizontalAlignment = leftAlignment,
-                    verticalArrangement = Arrangement.spacedBy(4.dp),
-                ) {
-                    leftLines.forEach { line ->
-                        Text(
-                            text = line.text,
-                            style = line.style,
-                            fontWeight = line.fontWeight,
-                            color = line.color ?: MaterialTheme.colorScheme.onSurfaceVariant,
-                            textAlign = if (shift.type == ShiftType.MARKET) TextAlign.Center else TextAlign.Start,
-                        )
-                    }
-                }
-
-                Column(
-                    modifier = Modifier.weight(0.62f),
-                    verticalArrangement = Arrangement.spacedBy(6.dp),
-                ) {
-                    primaryNames.forEachIndexed { index, name ->
-                        val marketStyle = MaterialTheme.typography.bodyMedium
-                        Text(
-                            text = name,
-                            style = if (shift.type == ShiftType.MARKET) {
-                                marketStyle
-                            } else if (index == 0) {
-                                MaterialTheme.typography.bodyLarge
-                            } else {
-                                MaterialTheme.typography.bodyMedium
-                            },
-                            fontWeight = if (shift.type == ShiftType.MARKET) FontWeight.Normal else if (index == 0) FontWeight.SemiBold else FontWeight.Normal,
-                            color = if (shift.type != ShiftType.MARKET && highlightedIndex == index) {
-                                MaterialTheme.colorScheme.primary
-                            } else {
-                                MaterialTheme.colorScheme.onSurface
-                            },
-                        )
-                    }
-                    if (shift.status != ShiftStatus.PLANNED) {
-                        Text(
-                            text = stringResource(shift.status.labelRes()),
-                            style = MaterialTheme.typography.labelMedium,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        )
-                    }
-                }
-            }
-            if (highlightedIndex != null && canRequestShiftSwap) {
-                ReguertaButton(
-                    label = stringResource(R.string.shift_swap_request_button_label),
-                    variant = ReguertaButtonVariant.SECONDARY,
-                    fullWidth = false,
-                    onClick = { onRequestShiftSwap(shift.id) },
-                )
-            }
-        }
-    }
-}
-
-private enum class ShiftBoardSegment(@StringRes val labelRes: Int) {
-    DELIVERY(R.string.shifts_type_delivery),
-    MARKET(R.string.shifts_type_market),
-}
-
-private data class ShiftBoardLine(
-    val text: String,
-    val style: androidx.compose.ui.text.TextStyle,
-    val fontWeight: FontWeight = FontWeight.Normal,
-    val color: androidx.compose.ui.graphics.Color? = null,
-)
-
-@Composable
-private fun ShiftSwapRequestsCard(
-    requests: List<ShiftSwapRequest>,
-    dismissedRequestIds: Set<String>,
-    shifts: List<ShiftAssignment>,
-    deliveryCalendarOverrides: List<DeliveryCalendarOverride>,
-    members: List<Member>,
-    currentMemberId: String?,
-    selectedSegment: ShiftBoardSegment,
-    isUpdating: Boolean,
-    onAccept: (String, String) -> Unit,
-    onReject: (String, String) -> Unit,
-    onCancel: (String) -> Unit,
-    onConfirm: (String, String) -> Unit,
-    onDismissRequest: (String) -> Unit,
-) {
-    val relevantRequests = remember(requests, shifts, selectedSegment) {
-        requests.filter { request ->
-            shifts.firstOrNull { it.id == request.requestedShiftId }?.type == selectedSegment.toShiftType()
-        }
-    }
-
-    Card {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp),
-        ) {
-            Text(
-                text = stringResource(R.string.shift_swap_requests_title),
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.SemiBold,
-            )
-            Text(
-                text = stringResource(R.string.shift_swap_requests_subtitle),
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-
-            if (relevantRequests.isEmpty() || currentMemberId == null) {
-                Text(
-                    text = stringResource(R.string.shift_swap_requests_empty),
-                    style = MaterialTheme.typography.bodyMedium,
-                )
-                return@Column
-            }
-
-            val incoming = relevantRequests.flatMap { request ->
-                request.candidates
-                    .filter { it.userId == currentMemberId }
-                    .filter { candidate ->
-                        request.status == ShiftSwapRequestStatus.OPEN &&
-                            request.responses.none { response ->
-                                response.userId == candidate.userId && response.shiftId == candidate.shiftId
-                            }
-                    }
-                    .map { candidate -> request to candidate }
-            }
-            val requesterOpen = relevantRequests.filter { it.requesterUserId == currentMemberId && it.status == ShiftSwapRequestStatus.OPEN }
-            val availableResponses = requesterOpen.flatMap { request ->
-                request.availableResponses().mapNotNull { response ->
-                    val candidate = request.candidates.firstOrNull { it.userId == response.userId && it.shiftId == response.shiftId }
-                    candidate?.let { Triple(request, it, response) }
-                }
-            }
-            val waiting = requesterOpen.filter { request -> request.availableResponses().isEmpty() }
-            val history = relevantRequests.filter { request ->
-                request.status != ShiftSwapRequestStatus.OPEN &&
-                    request.id !in dismissedRequestIds
-            }
-
-            if (incoming.isNotEmpty()) {
-                IncomingShiftSwapSection(
-                    title = stringResource(R.string.shift_swap_requests_incoming),
-                    pendingCandidates = incoming,
-                    shifts = shifts,
-                    deliveryCalendarOverrides = deliveryCalendarOverrides,
-                    members = members,
-                    isUpdating = isUpdating,
-                    onAccept = onAccept,
-                    onReject = onReject,
-                )
-            }
-
-            if (availableResponses.isNotEmpty()) {
-                RequesterResponsesSection(
-                    title = stringResource(R.string.shift_swap_requests_responses),
-                    responseOptions = availableResponses,
-                    shifts = shifts,
-                    deliveryCalendarOverrides = deliveryCalendarOverrides,
-                    members = members,
-                    isUpdating = isUpdating,
-                    onConfirm = onConfirm,
-                )
-            }
-
-            if (waiting.isNotEmpty()) {
-                WaitingShiftSwapSection(
-                    title = stringResource(R.string.shift_swap_requests_outgoing),
-                    requests = waiting,
-                    shifts = shifts,
-                    deliveryCalendarOverrides = deliveryCalendarOverrides,
-                    members = members,
-                    onCancel = onCancel,
-                )
-            }
-
-            if (history.isNotEmpty()) {
-                HistoryShiftSwapSection(
-                    title = stringResource(R.string.shift_swap_requests_history),
-                    requests = history,
-                    shifts = shifts,
-                    deliveryCalendarOverrides = deliveryCalendarOverrides,
-                    members = members,
-                    onDismissRequest = onDismissRequest,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun IncomingShiftSwapSection(
-    title: String,
-    pendingCandidates: List<Pair<ShiftSwapRequest, com.reguerta.user.domain.shifts.ShiftSwapCandidate>>,
-    shifts: List<ShiftAssignment>,
-    deliveryCalendarOverrides: List<DeliveryCalendarOverride>,
-    members: List<Member>,
-    isUpdating: Boolean,
-    onAccept: (String, String) -> Unit,
-    onReject: (String, String) -> Unit,
-) {
-    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-        Text(
-            text = title,
-            style = MaterialTheme.typography.labelLarge,
-            color = MaterialTheme.colorScheme.primary,
-            fontWeight = FontWeight.SemiBold,
-        )
-        pendingCandidates.sortedByDescending { it.first.requestedAtMillis }.forEach { (request, candidate) ->
-            val requestedShift = shifts.firstOrNull { it.id == request.requestedShiftId }
-            val candidateShift = shifts.firstOrNull { it.id == candidate.shiftId }
-            Card {
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(12.dp),
-                    verticalArrangement = Arrangement.spacedBy(8.dp),
-                ) {
-                    Text(
-                        text = stringResource(
-                            R.string.shift_swap_request_requested_by_format,
-                            members.displayNameFor(request.requesterUserId),
-                        ),
-                        style = MaterialTheme.typography.bodyMedium,
-                        fontWeight = FontWeight.SemiBold,
-                    )
-                    Text(
-                        text = stringResource(
-                            R.string.shift_swap_request_shift_format,
-                            requestedShift?.toShiftSwapDisplayLabel(request.requesterUserId, deliveryCalendarOverrides).orEmpty().ifBlank { request.requestedShiftId },
-                        ),
-                        style = MaterialTheme.typography.bodySmall,
-                    )
-                    Text(
-                        text = stringResource(
-                            R.string.shift_swap_request_offer_shift_format,
-                            candidateShift?.toShiftSwapDisplayLabel(candidate.userId, deliveryCalendarOverrides).orEmpty().ifBlank { candidate.shiftId },
-                        ),
-                        style = MaterialTheme.typography.bodySmall,
-                    )
-                    if (request.reason.isNotBlank()) {
-                        Text(
-                            text = stringResource(R.string.shift_swap_request_reason_format, request.reason),
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        )
-                    }
-                    Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                        ReguertaButton(
-                            label = stringResource(R.string.shift_swap_request_accept_short),
-                            variant = ReguertaButtonVariant.PRIMARY,
-                            fullWidth = false,
-                            enabled = !isUpdating,
-                            onClick = { onAccept(request.id, candidate.shiftId) },
-                        )
-                        ReguertaButton(
-                            label = stringResource(R.string.shift_swap_request_reject_short),
-                            variant = ReguertaButtonVariant.SECONDARY,
-                            fullWidth = false,
-                            enabled = !isUpdating,
-                            onClick = { onReject(request.id, candidate.shiftId) },
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun RequesterResponsesSection(
-    title: String,
-    responseOptions: List<Triple<ShiftSwapRequest, com.reguerta.user.domain.shifts.ShiftSwapCandidate, com.reguerta.user.domain.shifts.ShiftSwapResponse>>,
-    shifts: List<ShiftAssignment>,
-    deliveryCalendarOverrides: List<DeliveryCalendarOverride>,
-    members: List<Member>,
-    isUpdating: Boolean,
-    onConfirm: (String, String) -> Unit,
-) {
-    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-        Text(
-            text = title,
-            style = MaterialTheme.typography.labelLarge,
-            color = MaterialTheme.colorScheme.primary,
-            fontWeight = FontWeight.SemiBold,
-        )
-        responseOptions.sortedByDescending { it.first.requestedAtMillis }.forEach { (request, candidate, _) ->
-            val requestedShift = shifts.firstOrNull { it.id == request.requestedShiftId }
-            val candidateShift = shifts.firstOrNull { it.id == candidate.shiftId }
-            Card {
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(12.dp),
-                    verticalArrangement = Arrangement.spacedBy(8.dp),
-                ) {
-                    Text(
-                        text = members.displayNameFor(candidate.userId),
-                        style = MaterialTheme.typography.bodyMedium,
-                        fontWeight = FontWeight.SemiBold,
-                    )
-                    Text(
-                        text = stringResource(
-                            R.string.shift_swap_request_confirm_before_after_format,
-                            requestedShift?.toShiftSwapDisplayLabel(request.requesterUserId, deliveryCalendarOverrides).orEmpty().ifBlank { request.requestedShiftId },
-                            candidateShift?.toShiftSwapDisplayLabel(candidate.userId, deliveryCalendarOverrides).orEmpty().ifBlank { candidate.shiftId },
-                        ),
-                        style = MaterialTheme.typography.bodySmall,
-                    )
-                    ReguertaButton(
-                        label = stringResource(R.string.shift_swap_request_confirm),
-                        variant = ReguertaButtonVariant.PRIMARY,
-                        fullWidth = false,
-                        enabled = !isUpdating,
-                        onClick = { onConfirm(request.id, candidate.shiftId) },
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun WaitingShiftSwapSection(
-    title: String,
-    requests: List<ShiftSwapRequest>,
-    shifts: List<ShiftAssignment>,
-    deliveryCalendarOverrides: List<DeliveryCalendarOverride>,
-    members: List<Member>,
-    onCancel: (String) -> Unit,
-) {
-    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-        Text(
-            text = title,
-            style = MaterialTheme.typography.labelLarge,
-            color = MaterialTheme.colorScheme.primary,
-            fontWeight = FontWeight.SemiBold,
-        )
-        requests.sortedByDescending { it.requestedAtMillis }.forEach { request ->
-            val requestedShift = shifts.firstOrNull { it.id == request.requestedShiftId }
-            Card {
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(12.dp),
-                    verticalArrangement = Arrangement.spacedBy(8.dp),
-                ) {
-                    Text(
-                        text = requestedShift?.toShiftSwapDisplayLabel(request.requesterUserId, deliveryCalendarOverrides) ?: request.requestedShiftId,
-                        style = MaterialTheme.typography.bodyMedium,
-                        fontWeight = FontWeight.SemiBold,
-                    )
-                    Text(
-                        text = stringResource(R.string.shift_swap_request_waiting_multiple_format, request.candidates.map { members.displayNameFor(it.userId) }.distinct().size),
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    )
-                    ReguertaButton(
-                        label = stringResource(R.string.shift_swap_request_cancel),
-                        variant = ReguertaButtonVariant.SECONDARY,
-                        fullWidth = false,
-                        onClick = { onCancel(request.id) },
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun HistoryShiftSwapSection(
-    title: String,
-    requests: List<ShiftSwapRequest>,
-    shifts: List<ShiftAssignment>,
-    deliveryCalendarOverrides: List<DeliveryCalendarOverride>,
-    members: List<Member>,
-    onDismissRequest: (String) -> Unit,
-) {
-    Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-        Text(
-            text = title,
-            style = MaterialTheme.typography.labelLarge,
-            color = MaterialTheme.colorScheme.primary,
-            fontWeight = FontWeight.SemiBold,
-        )
-        requests.sortedByDescending { it.requestedAtMillis }.forEach { request ->
-            ShiftSwapRequestHistoryItem(
-                request = request,
-                shift = shifts.firstOrNull { it.id == request.requestedShiftId },
-                deliveryCalendarOverrides = deliveryCalendarOverrides,
-                members = members,
-                onDismiss = onDismissRequest,
-            )
-        }
-    }
-}
-
-@Composable
-private fun ShiftSwapRequestHistoryItem(
-    request: ShiftSwapRequest,
-    shift: ShiftAssignment?,
-    deliveryCalendarOverrides: List<DeliveryCalendarOverride>,
-    members: List<Member>,
-    onDismiss: (String) -> Unit,
-) {
-    Card {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(12.dp),
-            verticalArrangement = Arrangement.spacedBy(8.dp),
-        ) {
-            Text(
-                text = shift?.toShiftSwapDisplayLabel(request.requesterUserId, deliveryCalendarOverrides) ?: request.requestedShiftId,
-                style = MaterialTheme.typography.bodyMedium,
-                fontWeight = FontWeight.SemiBold,
-            )
-            Text(
-                text = stringResource(
-                    R.string.shift_swap_request_requested_by_format,
-                    members.displayNameFor(request.requesterUserId),
-                ),
-                style = MaterialTheme.typography.bodySmall,
-            )
-            Text(
-                text = stringResource(
-                    R.string.shift_swap_request_reason_format,
-                    request.reason.ifBlank { stringResource(R.string.shift_swap_request_reason_empty) },
-                ),
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-            Text(
-                text = stringResource(request.status.labelRes()),
-                style = MaterialTheme.typography.labelMedium,
-                color = MaterialTheme.colorScheme.primary,
-            )
-            request.selectedCandidateUserId?.let { selectedUserId ->
-                Text(
-                    text = stringResource(
-                        R.string.shift_swap_request_selected_candidate_format,
-                        members.displayNameFor(selectedUserId),
-                    ),
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-            }
-            if (request.status == ShiftSwapRequestStatus.APPLIED) {
-                ReguertaFlatButton(
-                    label = stringResource(R.string.shift_swap_request_acknowledge),
-                    onClick = { onDismiss(request.id) },
-                    modifier = Modifier.fillMaxWidth(),
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun ShiftSwapRequestRoute(
-    draft: ShiftSwapDraft,
-    shifts: List<ShiftAssignment>,
-    deliveryCalendarOverrides: List<DeliveryCalendarOverride>,
-    members: List<Member>,
-    isSaving: Boolean,
-    onDraftChanged: (ShiftSwapDraft) -> Unit,
-    onCancel: () -> Unit,
-    onSave: () -> Unit,
-) {
-    val focusManager = LocalFocusManager.current
-    val shift = shifts.firstOrNull { it.id == draft.shiftId }
-    Card {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp),
-        ) {
-            Text(
-                text = stringResource(R.string.shift_swap_request_screen_title),
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.SemiBold,
-            )
-            Text(
-                text = stringResource(R.string.shift_swap_request_screen_subtitle),
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-            Text(
-                text = stringResource(
-                    R.string.shift_swap_request_shift_format,
-                    shift?.let {
-                        it.toShiftSwapDisplayLabel(
-                            it.assignedUserIds.firstOrNull() ?: it.helperUserId,
-                            deliveryCalendarOverrides,
-                        )
-                    }.orEmpty().ifBlank { draft.shiftId },
-                ),
-                style = MaterialTheme.typography.bodyMedium,
-            )
-            Text(
-                text = stringResource(
-                    R.string.shift_swap_request_broadcast_scope_format,
-                    when (shift?.type) {
-                        ShiftType.MARKET -> stringResource(R.string.shifts_type_market)
-                        else -> stringResource(R.string.shifts_type_delivery)
-                    },
-                ),
-                style = MaterialTheme.typography.bodySmall,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-            OutlinedTextField(
-                value = draft.reason,
-                onValueChange = { onDraftChanged(draft.copy(reason = it)) },
-                modifier = Modifier.fillMaxWidth(),
-                label = { Text(stringResource(R.string.shift_swap_request_reason_label)) },
-                placeholder = { Text(stringResource(R.string.shift_swap_request_reason_placeholder)) },
-                minLines = 4,
-            )
-            Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                ReguertaButton(
-                    label = stringResource(
-                        if (isSaving) {
-                            R.string.shift_swap_request_saving
-                        } else {
-                            R.string.shift_swap_request_save
-                        },
-                    ),
-                    variant = ReguertaButtonVariant.PRIMARY,
-                    fullWidth = false,
-                    loading = isSaving,
-                    enabled = !isSaving && draft.shiftId.isNotBlank(),
-                    onClick = {
-                        focusManager.clearFocus(force = true)
-                        onSave()
-                    },
-                )
-                ReguertaButton(
-                    label = stringResource(R.string.common_action_back),
-                    variant = ReguertaButtonVariant.SECONDARY,
-                    fullWidth = false,
-                    enabled = !isSaving,
-                    onClick = {
-                        focusManager.clearFocus(force = true)
-                        onCancel()
-                    },
-                )
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun ProductsRoute(
-    currentMember: Member?,
-    products: List<Product>,
-    draft: ProductDraft,
-    editingProductId: String?,
-    isLoading: Boolean,
-    isSaving: Boolean,
-    isUpdatingProducerCatalogVisibility: Boolean,
-    onRefresh: () -> Unit,
-    onDraftChanged: (ProductDraft) -> Unit,
-    onCreateProduct: () -> Unit,
-    onEditProduct: (String) -> Unit,
-    onCancelEditor: () -> Unit,
-    onSaveProduct: () -> Unit,
-    onArchiveProduct: (String, onSuccess: () -> Unit) -> Unit,
-    onSetProducerCatalogVisibility: (Boolean, onSuccess: () -> Unit) -> Unit,
-) {
-    val focusManager = LocalFocusManager.current
-    val activeProducts = remember(products) { products.filterNot { it.archived } }
-    val archivedProducts = remember(products) { products.filter { it.archived } }
-    val isEditing = editingProductId != null
-    val canManageEcoBasket = currentMember?.isProducer == true
-    val canManageCommonPurchase = currentMember?.isCommonPurchaseManager == true && currentMember?.isProducer != true
-    var commonPurchaseExpanded by rememberSaveable { mutableStateOf(false) }
-    var pendingCatalogVisibility by rememberSaveable { mutableStateOf<Boolean?>(null) }
-
-    if (isEditing) {
-        Card {
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp),
-                verticalArrangement = Arrangement.spacedBy(12.dp),
-            ) {
-                Text(
-                    text = stringResource(
-                        if (editingProductId.isNullOrBlank()) {
-                            R.string.products_editor_title_create
-                        } else {
-                            R.string.products_editor_title_edit
-                        },
-                    ),
-                    style = MaterialTheme.typography.titleMedium,
-                    fontWeight = FontWeight.SemiBold,
-                )
-                Box(
-                    modifier = Modifier
-                        .size(112.dp)
-                        .clip(RoundedCornerShape(24.dp))
-                        .background(MaterialTheme.colorScheme.surfaceVariant),
-                    contentAlignment = Alignment.Center,
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.Image,
-                        contentDescription = null,
-                        tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                        modifier = Modifier.size(40.dp),
-                    )
-                }
-                Text(
-                    text = stringResource(R.string.products_editor_subtitle),
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-                OutlinedTextField(
-                    value = draft.name,
-                    onValueChange = { onDraftChanged(draft.copy(name = it)) },
-                    modifier = Modifier.fillMaxWidth(),
-                    label = { Text(stringResource(R.string.products_field_name)) },
-                )
-                OutlinedTextField(
-                    value = draft.description,
-                    onValueChange = { onDraftChanged(draft.copy(description = it)) },
-                    modifier = Modifier.fillMaxWidth(),
-                    label = { Text(stringResource(R.string.products_field_description)) },
-                    minLines = 3,
-                )
-                Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                    OutlinedTextField(
-                        value = draft.unitQty,
-                        onValueChange = { onDraftChanged(draft.copy(unitQty = it)) },
-                        modifier = Modifier.weight(1f),
-                        label = { Text(stringResource(R.string.products_field_unit_qty)) },
-                    )
-                    OutlinedTextField(
-                        value = draft.packContainerQty,
-                        onValueChange = { onDraftChanged(draft.copy(packContainerQty = it)) },
-                        modifier = Modifier.weight(1f),
-                        label = { Text(stringResource(R.string.products_field_pack_qty)) },
-                    )
-                }
-                Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                    OutlinedTextField(
-                        value = draft.packContainerName,
-                        onValueChange = { onDraftChanged(draft.copy(packContainerName = it)) },
-                        modifier = Modifier.weight(1f),
-                        label = { Text(stringResource(R.string.products_field_pack_name)) },
-                    )
-                    OutlinedTextField(
-                        value = draft.unitName,
-                        onValueChange = { onDraftChanged(draft.copy(unitName = it)) },
-                        modifier = Modifier.weight(1f),
-                        label = { Text(stringResource(R.string.products_field_unit_name)) },
-                    )
-                }
-                Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                    OutlinedTextField(
-                        value = draft.packContainerPlural,
-                        onValueChange = { onDraftChanged(draft.copy(packContainerPlural = it)) },
-                        modifier = Modifier.weight(1f),
-                        label = { Text(stringResource(R.string.products_field_pack_plural)) },
-                    )
-                    OutlinedTextField(
-                        value = draft.unitPlural,
-                        onValueChange = { onDraftChanged(draft.copy(unitPlural = it)) },
-                        modifier = Modifier.weight(1f),
-                        label = { Text(stringResource(R.string.products_field_unit_plural)) },
-                    )
-                }
-                Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                    OutlinedTextField(
-                        value = draft.price,
-                        onValueChange = { onDraftChanged(draft.copy(price = it)) },
-                        modifier = Modifier.weight(1f),
-                        label = { Text(stringResource(R.string.products_field_price)) },
-                    )
-                    OutlinedTextField(
-                        value = draft.stockQty,
-                        onValueChange = { onDraftChanged(draft.copy(stockQty = it)) },
-                        modifier = Modifier.weight(1f),
-                        label = { Text(stringResource(R.string.products_field_stock_qty)) },
-                        enabled = draft.stockMode == ProductStockMode.FINITE,
-                    )
-                }
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.spacedBy(12.dp),
-                    verticalAlignment = Alignment.CenterVertically,
-                ) {
-                    Text(
-                        text = stringResource(R.string.products_field_available),
-                        modifier = Modifier.weight(1f),
-                    )
-                    Switch(
-                        checked = draft.isAvailable,
-                        onCheckedChange = { onDraftChanged(draft.copy(isAvailable = it)) },
-                    )
-                }
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.spacedBy(12.dp),
-                    verticalAlignment = Alignment.CenterVertically,
-                ) {
-                    Text(
-                        text = stringResource(R.string.products_field_infinite_stock),
-                        modifier = Modifier.weight(1f),
-                    )
-                    Switch(
-                        checked = draft.stockMode == ProductStockMode.INFINITE,
-                        onCheckedChange = {
-                            onDraftChanged(
-                                draft.copy(
-                                    stockMode = if (it) ProductStockMode.INFINITE else ProductStockMode.FINITE,
-                                    stockQty = if (it) "" else draft.stockQty,
-                                ),
-                            )
-                        },
-                    )
-                }
-                if (canManageEcoBasket) {
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.spacedBy(12.dp),
-                        verticalAlignment = Alignment.CenterVertically,
-                    ) {
-                        Text(
-                            text = stringResource(R.string.products_field_eco_basket),
-                            modifier = Modifier.weight(1f),
-                        )
-                        Switch(
-                            checked = draft.isEcoBasket,
-                            onCheckedChange = { onDraftChanged(draft.copy(isEcoBasket = it)) },
-                        )
-                    }
-                }
-                if (canManageCommonPurchase) {
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.spacedBy(12.dp),
-                        verticalAlignment = Alignment.CenterVertically,
-                    ) {
-                        Text(
-                            text = stringResource(R.string.products_field_common_purchase),
-                            modifier = Modifier.weight(1f),
-                        )
-                        Switch(
-                            checked = draft.isCommonPurchase,
-                            onCheckedChange = {
-                                onDraftChanged(
-                                    draft.copy(
-                                        isCommonPurchase = it,
-                                        commonPurchaseType = if (it) {
-                                            draft.commonPurchaseType ?: CommonPurchaseType.SPOT
-                                        } else {
-                                            null
-                                        },
-                                    ),
-                                )
-                            },
-                        )
-                    }
-                    if (draft.isCommonPurchase) {
-                        ExposedDropdownMenuBox(
-                            expanded = commonPurchaseExpanded,
-                            onExpandedChange = { commonPurchaseExpanded = !commonPurchaseExpanded },
-                        ) {
-                            OutlinedTextField(
-                                value = stringResource(
-                                    when (draft.commonPurchaseType ?: CommonPurchaseType.SPOT) {
-                                        CommonPurchaseType.SEASONAL -> R.string.products_common_purchase_type_seasonal
-                                        CommonPurchaseType.SPOT -> R.string.products_common_purchase_type_spot
-                                    },
-                                ),
-                                onValueChange = {},
-                                modifier = Modifier
-                                    .menuAnchor(ExposedDropdownMenuAnchorType.PrimaryNotEditable)
-                                    .fillMaxWidth(),
-                                label = { Text(stringResource(R.string.products_field_common_purchase_type)) },
-                                readOnly = true,
-                                trailingIcon = {
-                                    ExposedDropdownMenuDefaults.TrailingIcon(expanded = commonPurchaseExpanded)
-                                },
-                            )
-                            ExposedDropdownMenu(
-                                expanded = commonPurchaseExpanded,
-                                onDismissRequest = { commonPurchaseExpanded = false },
-                            ) {
-                                DropdownMenuItem(
-                                    text = { Text(stringResource(R.string.products_common_purchase_type_spot)) },
-                                    onClick = {
-                                        onDraftChanged(draft.copy(commonPurchaseType = CommonPurchaseType.SPOT))
-                                        commonPurchaseExpanded = false
-                                    },
-                                )
-                                DropdownMenuItem(
-                                    text = { Text(stringResource(R.string.products_common_purchase_type_seasonal)) },
-                                    onClick = {
-                                        onDraftChanged(draft.copy(commonPurchaseType = CommonPurchaseType.SEASONAL))
-                                        commonPurchaseExpanded = false
-                                    },
-                                )
-                            }
-                        }
-                    }
-                }
-                Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                    ReguertaButton(
-                        label = stringResource(
-                            if (isSaving) R.string.products_save_action_saving else R.string.products_save_action
-                        ),
-                        variant = ReguertaButtonVariant.PRIMARY,
-                        fullWidth = false,
-                        loading = isSaving,
-                        enabled = !isSaving,
-                        onClick = {
-                            focusManager.clearFocus(force = true)
-                            onSaveProduct()
-                        },
-                    )
-                    ReguertaButton(
-                        label = stringResource(R.string.common_action_back),
-                        variant = ReguertaButtonVariant.SECONDARY,
-                        fullWidth = false,
-                        enabled = !isSaving,
-                        onClick = {
-                            focusManager.clearFocus(force = true)
-                            onCancelEditor()
-                        },
-                    )
-                }
-            }
-        }
-        return
-    }
-
-    Box(
-        modifier = Modifier.fillMaxWidth(),
-    ) {
-        Column(
-            verticalArrangement = Arrangement.spacedBy(16.dp),
-        ) {
-            Card {
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(16.dp),
-                    verticalArrangement = Arrangement.spacedBy(10.dp),
-                ) {
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.spacedBy(12.dp),
-                        verticalAlignment = Alignment.CenterVertically,
-                    ) {
-                        Text(
-                            text = stringResource(R.string.products_catalog_title),
-                            modifier = Modifier.weight(1f),
-                            style = MaterialTheme.typography.titleMedium,
-                            fontWeight = FontWeight.SemiBold,
-                        )
-                        if (currentMember?.isProducer == true) {
-                            Button(
-                                onClick = {
-                                    pendingCatalogVisibility = !currentMember.producerCatalogEnabled
-                                },
-                                enabled = !isUpdatingProducerCatalogVisibility,
-                                shape = RoundedCornerShape(18.dp),
-                                colors = ButtonDefaults.buttonColors(
-                                    containerColor = if (currentMember.producerCatalogEnabled) {
-                                        Color(0xFFF08D43)
-                                    } else {
-                                        MaterialTheme.colorScheme.primary
-                                    },
-                                    contentColor = MaterialTheme.colorScheme.onPrimary,
-                                    disabledContainerColor = MaterialTheme.colorScheme.surfaceVariant,
-                                    disabledContentColor = MaterialTheme.colorScheme.onSurfaceVariant,
-                                ),
-                                contentPadding = androidx.compose.foundation.layout.PaddingValues(
-                                    horizontal = 14.dp,
-                                    vertical = 8.dp,
-                                ),
-                            ) {
-                                Text(
-                                    text = stringResource(
-                                        if (currentMember.producerCatalogEnabled) {
-                                            R.string.products_visibility_disable_compact_action
-                                        } else {
-                                            R.string.products_visibility_enable_compact_action
-                                        },
-                                    ),
-                                    textAlign = TextAlign.Center,
-                                    style = MaterialTheme.typography.labelMedium,
-                                    lineHeight = 16.sp,
-                                )
-                            }
-                        }
-                    }
-                    ReguertaButton(
-                        label = stringResource(R.string.products_refresh_action),
-                        variant = ReguertaButtonVariant.TEXT,
-                        fullWidth = false,
-                        onClick = onRefresh,
-                    )
-                }
-            }
-
-            if (isLoading) {
-                Card {
-                    Text(
-                        text = stringResource(R.string.products_loading),
-                        modifier = Modifier.padding(16.dp),
-                    )
-                }
-            } else {
-                if (activeProducts.isEmpty()) {
-                    Card {
-                        Text(
-                            text = stringResource(R.string.products_empty_state),
-                            modifier = Modifier.padding(16.dp),
-                        )
-                    }
-                } else {
-                    activeProducts.forEach { product ->
-                        ProductListItem(
-                            product = product,
-                            onEdit = { onEditProduct(product.id) },
-                            onArchive = { onArchiveProduct(product.id) { } },
-                        )
-                    }
-                }
-
-                if (archivedProducts.isNotEmpty()) {
-                    Text(
-                        text = stringResource(R.string.products_archived_section_title),
-                        style = MaterialTheme.typography.labelLarge,
-                        color = MaterialTheme.colorScheme.primary,
-                        fontWeight = FontWeight.SemiBold,
-                    )
-                    archivedProducts.forEach { product ->
-                        ProductListItem(
-                            product = product,
-                            onEdit = { onEditProduct(product.id) },
-                            onArchive = {},
-                        )
-                    }
-                }
-            }
-
-            Spacer(modifier = Modifier.height(96.dp))
-        }
-
-        ReguertaButton(
-            label = stringResource(R.string.products_create_action),
-            variant = ReguertaButtonVariant.PRIMARY,
-            modifier = Modifier
-                .align(Alignment.BottomCenter)
-                .fillMaxWidth()
-                .navigationBarsPadding(),
-            onClick = onCreateProduct,
-        )
-    }
-
-    pendingCatalogVisibility?.let { targetEnabled ->
-        AlertDialog(
-            onDismissRequest = { pendingCatalogVisibility = null },
-            title = {
-                Text(
-                    text = stringResource(
-                        if (targetEnabled) {
-                            R.string.products_visibility_enable_dialog_title
-                        } else {
-                            R.string.products_visibility_disable_dialog_title
-                        },
-                    ),
-                )
-            },
-            text = {
-                Text(
-                    text = stringResource(
-                        if (targetEnabled) {
-                            R.string.products_visibility_enable_dialog_body
-                        } else {
-                            R.string.products_visibility_disable_dialog_body
-                        },
-                    ),
-                )
-            },
-            confirmButton = {
-                TextButton(
-                    onClick = {
-                        onSetProducerCatalogVisibility(targetEnabled) {
-                            pendingCatalogVisibility = null
-                        }
-                    },
-                    enabled = !isUpdatingProducerCatalogVisibility,
-                ) {
-                    Text(stringResource(R.string.common_action_confirm))
-                }
-            },
-            dismissButton = {
-                TextButton(
-                    onClick = { pendingCatalogVisibility = null },
-                    enabled = !isUpdatingProducerCatalogVisibility,
-                ) {
-                    Text(stringResource(R.string.common_action_cancel))
-                }
-            },
-        )
-    }
-}
-
-@Composable
-private fun ProductListItem(
-    product: Product,
-    onEdit: () -> Unit,
-    onArchive: () -> Unit,
-) {
-    Card {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp),
-            horizontalArrangement = Arrangement.spacedBy(16.dp),
-            verticalAlignment = Alignment.Top,
-        ) {
-            Box(
-                modifier = Modifier
-                    .size(96.dp)
-                    .clip(RoundedCornerShape(20.dp))
-                    .background(MaterialTheme.colorScheme.surfaceVariant),
-                contentAlignment = Alignment.Center,
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Image,
-                    contentDescription = null,
-                    tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                    modifier = Modifier.size(32.dp),
-                )
-            }
-            Column(
-                modifier = Modifier.weight(1f),
-                verticalArrangement = Arrangement.spacedBy(8.dp),
-            ) {
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalAlignment = Alignment.Top,
-                ) {
-                    Column(
-                        modifier = Modifier.weight(1f),
-                        verticalArrangement = Arrangement.spacedBy(6.dp),
-                    ) {
-                        Text(
-                            text = product.name,
-                            style = MaterialTheme.typography.titleMedium,
-                            fontWeight = FontWeight.SemiBold,
-                        )
-                        Text(
-                            text = product.description.ifBlank { "Sin descripcion." },
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        )
-                    }
-                    Row {
-                        IconButton(onClick = onEdit) {
-                            Icon(
-                                imageVector = Icons.Default.Edit,
-                                contentDescription = stringResource(R.string.products_edit_action),
-                            )
-                        }
-                        if (!product.archived) {
-                            IconButton(onClick = onArchive) {
-                                Icon(
-                                    imageVector = Icons.Default.Delete,
-                                    contentDescription = stringResource(R.string.products_archive_action),
-                                )
-                            }
-                        }
-                    }
-                }
-                Text(
-                    text = "${product.price.toUiDecimal()} €",
-                    style = MaterialTheme.typography.headlineSmall,
-                    fontWeight = FontWeight.SemiBold,
-                )
-                Text(
-                    text = when {
-                        product.archived -> stringResource(R.string.products_status_archived)
-                        product.stockMode == ProductStockMode.INFINITE -> stringResource(R.string.products_status_infinite_stock)
-                        else -> stringResource(
-                            R.string.products_status_finite_stock_format,
-                            product.stockQty?.toUiDecimal().orEmpty().ifBlank { "0" },
-                        )
-                    },
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-                Text(
-                    text = "${product.unitQty.toUiDecimal()} ${product.unitAbbreviation ?: product.unitName}",
-                    style = MaterialTheme.typography.bodyMedium,
-                )
-                if (product.packContainerName != null) {
-                    Text(
-                        text = product.packContainerName,
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun LatestNewsCard(
-    news: List<NewsArticle>,
-    onViewAll: () -> Unit,
-) {
-    Card {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(10.dp),
-        ) {
-            Text(
-                text = stringResource(R.string.home_shell_news_title),
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.SemiBold,
-            )
-            if (news.isEmpty()) {
-                Text(
-                    text = stringResource(R.string.news_empty_state),
-                    style = MaterialTheme.typography.bodyMedium,
-                )
-            } else {
-                news.forEach { article ->
-                    Column(verticalArrangement = Arrangement.spacedBy(4.dp)) {
-                        Text(
-                            text = article.title,
-                            style = MaterialTheme.typography.bodyLarge,
-                            fontWeight = FontWeight.SemiBold,
-                        )
-                        Text(
-                            text = article.body,
-                            style = MaterialTheme.typography.bodySmall,
-                            maxLines = 3,
-                        )
-                    }
-                }
-            }
-            ReguertaFlatButton(
-                label = stringResource(R.string.news_view_all),
-                onClick = onViewAll,
-                modifier = Modifier.fillMaxWidth(),
-            )
-        }
-    }
-}
-
-@Composable
-private fun NewsFeedRoute(
-    articles: List<NewsArticle>,
-    isLoading: Boolean,
-    isAdmin: Boolean,
-    onRefresh: () -> Unit,
-    onCreateNews: () -> Unit,
-    onEditNews: (String) -> Unit,
-    onRequestDeleteNews: (String) -> Unit,
-) {
-    Column(verticalArrangement = Arrangement.spacedBy(16.dp)) {
-        Card {
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp),
-                verticalArrangement = Arrangement.spacedBy(12.dp),
-            ) {
-                Text(
-                    text = stringResource(R.string.home_shell_news_title),
-                    style = MaterialTheme.typography.titleMedium,
-                    fontWeight = FontWeight.SemiBold,
-                )
-                Text(
-                    text = stringResource(R.string.news_list_subtitle),
-                    style = MaterialTheme.typography.bodyMedium,
-                )
-                if (isAdmin) {
-                    ReguertaFullButton(
-                        label = stringResource(R.string.news_create_action),
-                        onClick = onCreateNews,
-                        fullWidth = true,
-                    )
-                }
-                ReguertaFlatButton(
-                    label = stringResource(R.string.news_refresh_action),
-                    onClick = onRefresh,
-                    modifier = Modifier.fillMaxWidth(),
-                )
-            }
-        }
-
-        if (isLoading) {
-            Card {
-                Text(
-                    text = stringResource(R.string.news_loading),
-                    modifier = Modifier.padding(16.dp),
-                )
-            }
-        } else if (articles.isEmpty()) {
-            Card {
-                Text(
-                    text = stringResource(R.string.news_empty_state),
-                    modifier = Modifier.padding(16.dp),
-                )
-            }
-        } else {
-            articles.forEach { article ->
-                Card {
-                    Column(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(16.dp),
-                        verticalArrangement = Arrangement.spacedBy(8.dp),
-                    ) {
-                        Text(
-                            text = article.title,
-                            style = MaterialTheme.typography.titleMedium,
-                            fontWeight = FontWeight.SemiBold,
-                        )
-                        Text(
-                            text = stringResource(R.string.news_meta_format, article.publishedBy),
-                            style = MaterialTheme.typography.labelMedium,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        )
-                        if (!article.active) {
-                            Text(
-                                text = stringResource(R.string.news_inactive_badge),
-                                style = MaterialTheme.typography.labelMedium,
-                                color = MaterialTheme.colorScheme.primary,
-                            )
-                        }
-                        Text(
-                            text = article.body,
-                            style = MaterialTheme.typography.bodyMedium,
-                        )
-                        article.urlImage?.let { url ->
-                            Text(
-                                text = url,
-                                style = MaterialTheme.typography.bodySmall,
-                                color = MaterialTheme.colorScheme.primary,
-                            )
-                        }
-                        if (isAdmin) {
-                            Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                                ReguertaFlatButton(
-                                    label = stringResource(R.string.news_edit_action),
-                                    onClick = { onEditNews(article.id) },
-                                )
-                                TextButton(onClick = { onRequestDeleteNews(article.id) }) {
-                                    Text(stringResource(R.string.news_delete_action))
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun NewsEditorRoute(
-    draft: NewsDraft,
-    isSaving: Boolean,
-    isEditing: Boolean,
-    onDraftChanged: (NewsDraft) -> Unit,
-    onCancel: () -> Unit,
-    onSave: () -> Unit,
-) {
-    val focusManager = LocalFocusManager.current
-    Card {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .imePadding()
-                .navigationBarsPadding()
-                .padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp),
-        ) {
-            Text(
-                text = stringResource(
-                    if (isEditing) R.string.news_editor_title_edit else R.string.news_editor_title_create,
-                ),
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.SemiBold,
-            )
-            OutlinedTextField(
-                value = draft.title,
-                onValueChange = { onDraftChanged(draft.copy(title = it)) },
-                modifier = Modifier.fillMaxWidth(),
-                label = { Text(stringResource(R.string.news_field_title)) },
-                enabled = !isSaving,
-            )
-            OutlinedTextField(
-                value = draft.body,
-                onValueChange = { onDraftChanged(draft.copy(body = it)) },
-                modifier = Modifier.fillMaxWidth(),
-                label = { Text(stringResource(R.string.news_field_body)) },
-                minLines = 6,
-                enabled = !isSaving,
-            )
-            OutlinedTextField(
-                value = draft.urlImage,
-                onValueChange = { onDraftChanged(draft.copy(urlImage = it)) },
-                modifier = Modifier.fillMaxWidth(),
-                label = { Text(stringResource(R.string.news_field_url_image)) },
-                enabled = !isSaving,
-            )
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.SpaceBetween,
-            ) {
-                Text(text = stringResource(R.string.news_field_active))
-                Switch(
-                    checked = draft.active,
-                    onCheckedChange = { onDraftChanged(draft.copy(active = it)) },
-                    enabled = !isSaving,
-                )
-            }
-            ReguertaFullButton(
-                label = stringResource(
-                    if (isSaving) {
-                        R.string.news_save_action_saving
-                    } else if (isEditing) {
-                        R.string.news_save_action_update
-                    } else {
-                        R.string.news_save_action_create
-                    },
-                ),
-                onClick = {
-                    focusManager.clearFocus(force = true)
-                    onSave()
-                },
-                fullWidth = true,
-                enabled = !isSaving,
-            )
-            ReguertaFlatButton(
-                label = stringResource(R.string.common_action_back),
-                onClick = {
-                    focusManager.clearFocus(force = true)
-                    onCancel()
-                },
-                modifier = Modifier.fillMaxWidth(),
-                enabled = !isSaving,
-            )
-            Spacer(modifier = Modifier.height(64.dp))
-        }
-    }
-}
-
-@Composable
-private fun NotificationsFeedRoute(
-    notifications: List<NotificationEvent>,
-    isLoading: Boolean,
-    isAdmin: Boolean,
-    onRefresh: () -> Unit,
-    onCreateNotification: () -> Unit,
-) {
-    Column(verticalArrangement = Arrangement.spacedBy(16.dp)) {
-        Card {
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp),
-                verticalArrangement = Arrangement.spacedBy(12.dp),
-            ) {
-                Text(
-                    text = stringResource(R.string.home_shell_notifications),
-                    style = MaterialTheme.typography.titleMedium,
-                    fontWeight = FontWeight.SemiBold,
-                )
-                Text(
-                    text = stringResource(R.string.notifications_list_subtitle),
-                    style = MaterialTheme.typography.bodyMedium,
-                )
-                if (isAdmin) {
-                    ReguertaFullButton(
-                        label = stringResource(R.string.notifications_create_action),
-                        onClick = onCreateNotification,
-                        fullWidth = true,
-                    )
-                }
-                ReguertaFlatButton(
-                    label = stringResource(R.string.notifications_refresh_action),
-                    onClick = onRefresh,
-                    modifier = Modifier.fillMaxWidth(),
-                )
-            }
-        }
-
-        if (isLoading) {
-            Card {
-                Text(
-                    text = stringResource(R.string.notifications_loading),
-                    modifier = Modifier.padding(16.dp),
-                )
-            }
-        } else if (notifications.isEmpty()) {
-            Card {
-                Text(
-                    text = stringResource(R.string.notifications_empty_state),
-                    modifier = Modifier.padding(16.dp),
-                )
-            }
-        } else {
-            notifications.forEach { event ->
-                Card {
-                    Column(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(16.dp),
-                        verticalArrangement = Arrangement.spacedBy(8.dp),
-                    ) {
-                        Text(
-                            text = event.title,
-                            style = MaterialTheme.typography.titleMedium,
-                            fontWeight = FontWeight.SemiBold,
-                        )
-                        Text(
-                            text = stringResource(
-                                R.string.notifications_meta_format,
-                                event.sentAtMillis.toLocalizedDateTime(),
-                            ),
-                            style = MaterialTheme.typography.labelMedium,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        )
-                        Text(
-                            text = event.body,
-                            style = MaterialTheme.typography.bodyMedium,
-                        )
-                        Text(
-                            text = stringResource(event.audienceLabelRes()),
-                            style = MaterialTheme.typography.labelMedium,
-                            color = MaterialTheme.colorScheme.primary,
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun NotificationEditorRoute(
-    draft: NotificationDraft,
-    isSending: Boolean,
-    onDraftChanged: (NotificationDraft) -> Unit,
-    onCancel: () -> Unit,
-    onSend: () -> Unit,
-) {
-    val focusManager = LocalFocusManager.current
-    Card {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .imePadding()
-                .navigationBarsPadding()
-                .padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp),
-        ) {
-            Text(
-                text = stringResource(R.string.notifications_editor_title),
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.SemiBold,
-            )
-            Text(
-                text = stringResource(R.string.notifications_editor_subtitle),
-                style = MaterialTheme.typography.bodyMedium,
-            )
-            OutlinedTextField(
-                value = draft.title,
-                onValueChange = { onDraftChanged(draft.copy(title = it)) },
-                modifier = Modifier.fillMaxWidth(),
-                label = { Text(stringResource(R.string.notifications_field_title)) },
-                enabled = !isSending,
-            )
-            OutlinedTextField(
-                value = draft.body,
-                onValueChange = { onDraftChanged(draft.copy(body = it)) },
-                modifier = Modifier.fillMaxWidth(),
-                minLines = 5,
-                label = { Text(stringResource(R.string.notifications_field_body)) },
-                enabled = !isSending,
-            )
-            Text(
-                text = stringResource(R.string.notifications_field_audience),
-                style = MaterialTheme.typography.titleSmall,
-                fontWeight = FontWeight.SemiBold,
-            )
-            NotificationAudience.values().forEach { audience ->
-                ReguertaFlatButton(
-                    label = buildString {
-                        if (draft.audience == audience) append("• ")
-                        append(stringResource(audience.labelRes()))
-                    },
-                    onClick = { onDraftChanged(draft.copy(audience = audience)) },
-                    modifier = Modifier.fillMaxWidth(),
-                    enabled = !isSending,
-                )
-            }
-            ReguertaFullButton(
-                label = stringResource(
-                    if (isSending) {
-                        R.string.notifications_send_action_sending
-                    } else {
-                        R.string.notifications_send_action_send
-                    },
-                ),
-                onClick = {
-                    focusManager.clearFocus(force = true)
-                    onSend()
-                },
-                enabled = !isSending,
-                fullWidth = true,
-            )
-            ReguertaFlatButton(
-                label = stringResource(R.string.common_action_back),
-                onClick = {
-                    focusManager.clearFocus(force = true)
-                    onCancel()
-                },
-                modifier = Modifier.fillMaxWidth(),
-                enabled = !isSending,
-            )
-            Spacer(modifier = Modifier.height(48.dp))
-        }
-    }
-}
-
-@Composable
 private fun HomePlaceholderRoute(
     title: String,
     subtitle: String,
@@ -3663,481 +1579,6 @@ private fun HomePlaceholderRoute(
             )
         }
     }
-}
-
-@Composable
-private fun SettingsRoute(
-    currentMember: Member?,
-    authenticatedMember: Member?,
-    members: List<Member>,
-    shifts: List<ShiftAssignment>,
-    deliveryCalendarOverrides: List<DeliveryCalendarOverride>,
-    defaultDeliveryDayOfWeek: DeliveryWeekday?,
-    isLoadingDeliveryCalendar: Boolean,
-    isSavingDeliveryCalendar: Boolean,
-    isSubmittingShiftPlanningRequest: Boolean,
-    isDevelopImpersonationEnabled: Boolean,
-    onImpersonateMember: (String) -> Unit,
-    onClearImpersonation: () -> Unit,
-    onRefreshDeliveryCalendar: () -> Unit,
-    onSaveDeliveryCalendarOverride: (String, DeliveryWeekday, String, onSuccess: () -> Unit) -> Unit,
-    onDeleteDeliveryCalendarOverride: (String, onSuccess: () -> Unit) -> Unit,
-    onSubmitShiftPlanningRequest: (ShiftPlanningRequestType, onSuccess: () -> Unit) -> Unit,
-) {
-    var isImpersonationExpanded by rememberSaveable { mutableStateOf(false) }
-    Card {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp),
-        ) {
-            Text(
-                text = "Ajustes",
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.SemiBold,
-            )
-            Text(
-                text = "La impersonacion solo aparece en develop para probar flujos con otros socios sin salir de tu sesion real.",
-                style = MaterialTheme.typography.bodyMedium,
-            )
-            if (isDevelopImpersonationEnabled && currentMember != null && authenticatedMember != null) {
-                val isImpersonating = currentMember.id != authenticatedMember.id
-                Text(
-                    text = "Cuenta real: ${authenticatedMember.displayName}",
-                    style = MaterialTheme.typography.bodyMedium,
-                    fontWeight = FontWeight.Medium,
-                )
-                Text(
-                    text = if (isImpersonating) {
-                        "Viendo la app como: ${currentMember.displayName}"
-                    } else {
-                        "Ahora mismo estas usando tu propio perfil."
-                    },
-                    style = MaterialTheme.typography.bodyMedium,
-                )
-                if (isImpersonating) {
-                    ReguertaFlatButton(
-                        label = "Volver a mi perfil real",
-                        onClick = onClearImpersonation,
-                        modifier = Modifier.fillMaxWidth(),
-                    )
-                }
-                HorizontalDivider()
-                Text(
-                    text = "Impersonacion develop",
-                    style = MaterialTheme.typography.titleSmall,
-                    fontWeight = FontWeight.SemiBold,
-                )
-                ReguertaFlatButton(
-                    label = if (isImpersonationExpanded) {
-                        "Ocultar socios"
-                    } else {
-                        "Elegir socio"
-                    },
-                    onClick = { isImpersonationExpanded = !isImpersonationExpanded },
-                    modifier = Modifier.fillMaxWidth(),
-                )
-                if (isImpersonationExpanded) {
-                    members
-                        .filter { it.isActive }
-                        .sortedBy { it.displayName.lowercase(Locale.getDefault()) }
-                        .forEach { member ->
-                            val isSelected = member.id == currentMember.id
-                            ReguertaFlatButton(
-                                label = member.displayName,
-                                onClick = {
-                                    onImpersonateMember(member.id)
-                                    isImpersonationExpanded = false
-                                },
-                                modifier = Modifier.fillMaxWidth(),
-                                enabled = !isSelected,
-                            )
-                        }
-                }
-            }
-            if (currentMember?.isAdmin == true) {
-                HorizontalDivider()
-                AdminDeliveryCalendarSection(
-                    currentMember = currentMember,
-                    shifts = shifts,
-                    overrides = deliveryCalendarOverrides,
-                    defaultDeliveryDayOfWeek = defaultDeliveryDayOfWeek,
-                    isLoading = isLoadingDeliveryCalendar,
-                    isSaving = isSavingDeliveryCalendar,
-                    onRefresh = onRefreshDeliveryCalendar,
-                    onSaveOverride = onSaveDeliveryCalendarOverride,
-                    onDeleteOverride = onDeleteDeliveryCalendarOverride,
-                )
-                HorizontalDivider()
-                AdminShiftPlanningSection(
-                    isSubmitting = isSubmittingShiftPlanningRequest,
-                    onSubmit = onSubmitShiftPlanningRequest,
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun AdminDeliveryCalendarSection(
-    currentMember: Member,
-    shifts: List<ShiftAssignment>,
-    overrides: List<DeliveryCalendarOverride>,
-    defaultDeliveryDayOfWeek: DeliveryWeekday?,
-    isLoading: Boolean,
-    isSaving: Boolean,
-    onRefresh: () -> Unit,
-    onSaveOverride: (String, DeliveryWeekday, String, onSuccess: () -> Unit) -> Unit,
-    onDeleteOverride: (String, onSuccess: () -> Unit) -> Unit,
-) {
-    val futureWeeks = remember(shifts, overrides) {
-        shifts.filter { it.type == ShiftType.DELIVERY && it.effectiveDateMillis(overrides) > System.currentTimeMillis() }
-            .sortedBy { it.effectiveDateMillis(overrides) }
-            .distinctBy { it.dateMillis.toWeekKey() }
-    }
-    var isPickerVisible by rememberSaveable { mutableStateOf(false) }
-    var selectedWeekKey by rememberSaveable { mutableStateOf<String?>(null) }
-    Text(
-        text = "Calendario de reparto",
-        style = MaterialTheme.typography.titleSmall,
-        fontWeight = FontWeight.SemiBold,
-    )
-    Text(
-        text = "Gestiona excepciones por semana. Si quitas una excepcion, esa semana vuelve al dia por defecto del calendario.",
-        style = MaterialTheme.typography.bodyMedium,
-    )
-    Text(
-        text = "Dia por defecto: ${defaultDeliveryDayOfWeek?.toSpanishLabel() ?: "sin configurar"}",
-        style = MaterialTheme.typography.bodyMedium,
-        fontWeight = FontWeight.Medium,
-    )
-    if (isLoading) {
-        Text(
-            text = "Cargando calendario...",
-            style = MaterialTheme.typography.bodyMedium,
-        )
-    } else if (futureWeeks.isEmpty()) {
-        Text(
-            text = "No hay semanas de reparto futuras en los turnos cargados.",
-            style = MaterialTheme.typography.bodyMedium,
-        )
-    } else {
-        Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-            Button(onClick = { isPickerVisible = true }) {
-                Text("Cambiar dia de reparto")
-            }
-            TextButton(onClick = onRefresh) {
-                Text("Recargar")
-            }
-        }
-        Text(
-            text = "Primero eliges la semana a cambiar y despues editas solo esa excepcion.",
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-        )
-        if (isPickerVisible) {
-            DeliveryCalendarWeekPickerDialog(
-                futureWeeks = futureWeeks,
-                overrides = overrides,
-                onDismiss = { isPickerVisible = false },
-                onSelectWeek = { weekKey ->
-                    selectedWeekKey = weekKey
-                    isPickerVisible = false
-                },
-            )
-        }
-        selectedWeekKey?.let { weekKey ->
-            val selectedShift = futureWeeks.firstOrNull { it.dateMillis.toWeekKey() == weekKey }
-            if (selectedShift != null) {
-                DeliveryCalendarOverrideDialog(
-                    currentMember = currentMember,
-                    selectedShift = selectedShift,
-                    override = overrides.firstOrNull { it.weekKey == weekKey },
-                    defaultDeliveryDayOfWeek = defaultDeliveryDayOfWeek ?: DeliveryWeekday.WEDNESDAY,
-                    isSaving = isSaving,
-                    onDismiss = { selectedWeekKey = null },
-                    onSaveOverride = onSaveOverride,
-                    onDeleteOverride = onDeleteOverride,
-                )
-            } else {
-                selectedWeekKey = null
-            }
-        }
-    }
-}
-
-@Composable
-private fun AdminShiftPlanningSection(
-    isSubmitting: Boolean,
-    onSubmit: (ShiftPlanningRequestType, onSuccess: () -> Unit) -> Unit,
-) {
-    var pendingType by rememberSaveable { mutableStateOf<ShiftPlanningRequestType?>(null) }
-
-    Text(
-        text = "Planificacion de turnos",
-        style = MaterialTheme.typography.titleSmall,
-        fontWeight = FontWeight.SemiBold,
-    )
-    Text(
-        text = "Genera una temporada nueva con socios activos, escribe la hoja nueva y manda una notificacion a los socios asignados.",
-        style = MaterialTheme.typography.bodyMedium,
-    )
-    Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-        Button(
-            onClick = { pendingType = ShiftPlanningRequestType.DELIVERY },
-            enabled = !isSubmitting,
-        ) {
-            Text("Generar reparto")
-        }
-        Button(
-            onClick = { pendingType = ShiftPlanningRequestType.MARKET },
-            enabled = !isSubmitting,
-        ) {
-            Text("Generar mercado")
-        }
-    }
-    if (isSubmitting) {
-        Text(
-            text = "Enviando solicitud de planificacion...",
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-        )
-    }
-
-    pendingType?.let { type ->
-        val readableType = if (type == ShiftPlanningRequestType.DELIVERY) {
-            "reparto"
-        } else {
-            "mercado"
-        }
-        ReguertaDialog(
-            type = ReguertaDialogType.INFO,
-            title = "Generar turnos de $readableType",
-            message = "Se creara una planificacion nueva con socios activos, se escribira en la sheet de la temporada siguiente y se notificara a los socios asignados. Si vuelves a lanzarlo, se regenerara esa temporada.",
-            primaryAction = ReguertaDialogAction(
-                label = "Confirmar",
-                onClick = {
-                    onSubmit(type) {
-                        pendingType = null
-                    }
-                },
-            ),
-            secondaryAction = ReguertaDialogAction(
-                label = "Cancelar",
-                onClick = { pendingType = null },
-            ),
-            onDismissRequest = { pendingType = null },
-        )
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun DeliveryCalendarWeekPickerDialog(
-    futureWeeks: List<ShiftAssignment>,
-    overrides: List<DeliveryCalendarOverride>,
-    onDismiss: () -> Unit,
-    onSelectWeek: (String) -> Unit,
-) {
-    val initialSelection = futureWeeks.firstOrNull()?.dateMillis?.toWeekKey().orEmpty()
-    var selectedWeekKey by rememberSaveable(futureWeeks) { mutableStateOf(initialSelection) }
-    var isMenuExpanded by rememberSaveable { mutableStateOf(false) }
-    val selectedShift = futureWeeks.firstOrNull { it.dateMillis.toWeekKey() == selectedWeekKey }
-
-    ModalBottomSheet(
-        onDismissRequest = onDismiss,
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 24.dp, vertical = 12.dp)
-                .navigationBarsPadding()
-                .imePadding(),
-            verticalArrangement = Arrangement.spacedBy(16.dp),
-        ) {
-            Text(
-                text = "Elegir semana",
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.SemiBold,
-            )
-            Text(
-                text = "Selecciona un dia de reparto futuro con encargado.",
-                style = MaterialTheme.typography.bodyMedium,
-            )
-            ExposedDropdownMenuBox(
-                expanded = isMenuExpanded,
-                onExpandedChange = { isMenuExpanded = !isMenuExpanded },
-            ) {
-                OutlinedTextField(
-                    value = selectedShift?.let {
-                        "${it.dateMillis.toWeekKey()} · ${it.effectiveDateMillis(overrides).toLocalizedDateOnly()}"
-                    }.orEmpty(),
-                    onValueChange = {},
-                    readOnly = true,
-                    label = { Text("Semana de reparto") },
-                    trailingIcon = {
-                        ExposedDropdownMenuDefaults.TrailingIcon(expanded = isMenuExpanded)
-                    },
-                    modifier = Modifier
-                        .menuAnchor(ExposedDropdownMenuAnchorType.PrimaryNotEditable)
-                        .fillMaxWidth(),
-                )
-                ExposedDropdownMenu(
-                    expanded = isMenuExpanded,
-                    onDismissRequest = { isMenuExpanded = false },
-                ) {
-                    futureWeeks.forEach { shift ->
-                        val weekKey = shift.dateMillis.toWeekKey()
-                        DropdownMenuItem(
-                            text = {
-                                Text("${weekKey} · ${shift.effectiveDateMillis(overrides).toLocalizedDateOnly()}")
-                            },
-                            onClick = {
-                                selectedWeekKey = weekKey
-                                isMenuExpanded = false
-                            },
-                        )
-                    }
-                }
-            }
-            selectedShift?.let { shift ->
-                Card {
-                    Column(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(16.dp),
-                        verticalArrangement = Arrangement.spacedBy(6.dp),
-                    ) {
-                        Text(
-                            text = shift.dateMillis.toWeekKey(),
-                            style = MaterialTheme.typography.titleSmall,
-                            fontWeight = FontWeight.SemiBold,
-                        )
-                        Text(
-                            text = shift.effectiveDateMillis(overrides).toLocalizedDateOnly(),
-                            style = MaterialTheme.typography.bodyMedium,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        )
-                    }
-                }
-            }
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.End,
-            ) {
-                TextButton(onClick = onDismiss) {
-                    Text("Cerrar")
-                }
-                Button(
-                    onClick = { onSelectWeek(selectedWeekKey) },
-                    enabled = selectedWeekKey.isNotBlank(),
-                ) {
-                    Text("Elegir")
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun DeliveryCalendarOverrideDialog(
-    currentMember: Member,
-    selectedShift: ShiftAssignment,
-    override: DeliveryCalendarOverride?,
-    defaultDeliveryDayOfWeek: DeliveryWeekday,
-    isSaving: Boolean,
-    onDismiss: () -> Unit,
-    onSaveOverride: (String, DeliveryWeekday, String, onSuccess: () -> Unit) -> Unit,
-    onDeleteOverride: (String, onSuccess: () -> Unit) -> Unit,
-) {
-    val weekKey = selectedShift.dateMillis.toWeekKey()
-    var selectedWeekday by rememberSaveable(weekKey, override?.deliveryDateMillis) {
-        mutableStateOf(override?.deliveryDateMillis?.toDeliveryWeekday() ?: defaultDeliveryDayOfWeek)
-    }
-
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = {
-            Text(
-                text = "Cambiar dia de reparto",
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.SemiBold,
-            )
-        },
-        text = {
-            Column(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-                Text(
-                    text = weekKey,
-                    style = MaterialTheme.typography.bodyLarge,
-                    fontWeight = FontWeight.SemiBold,
-                )
-                Text(
-                    text = (override?.deliveryDateMillis ?: selectedShift.dateMillis).toLocalizedDateOnly(),
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-                Card {
-                    Column(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(12.dp),
-                        verticalArrangement = Arrangement.spacedBy(8.dp),
-                    ) {
-                        Text(
-                            text = if (override != null) {
-                                "Excepcion activa: ${override.deliveryDateMillis.toLocalizedDateOnly()}"
-                            } else {
-                                "Sin excepcion. Aplica el dia por defecto."
-                            },
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        )
-                        Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                            ReguertaFlatButton(
-                                label = "Anterior",
-                                onClick = { selectedWeekday = selectedWeekday.previous() },
-                                enabled = !isSaving,
-                            )
-                            Text(
-                                text = selectedWeekday.toSpanishLabel(),
-                                style = MaterialTheme.typography.bodyMedium,
-                                fontWeight = FontWeight.Medium,
-                                modifier = Modifier.align(Alignment.CenterVertically),
-                            )
-                            ReguertaFlatButton(
-                                label = "Siguiente",
-                                onClick = { selectedWeekday = selectedWeekday.next() },
-                                enabled = !isSaving,
-                            )
-                        }
-                        ReguertaButton(
-                            label = "Guardar excepcion",
-                            onClick = {
-                                onSaveOverride(weekKey, selectedWeekday, currentMember.id, onDismiss)
-                            },
-                            enabled = !isSaving,
-                            loading = isSaving,
-                            fullWidth = true,
-                        )
-                        if (override != null) {
-                            ReguertaFlatButton(
-                                label = "Quitar excepcion",
-                                onClick = { onDeleteOverride(weekKey, onDismiss) },
-                                modifier = Modifier.fillMaxWidth(),
-                                enabled = !isSaving,
-                            )
-                        }
-                    }
-                }
-            }
-        },
-        confirmButton = {
-            TextButton(onClick = onDismiss) {
-                Text("Cerrar")
-            }
-        },
-    )
 }
 
 @Composable
@@ -4532,26 +1973,26 @@ private fun ShiftType.labelRes(): Int = when (this) {
 }
 
 @StringRes
-private fun ShiftStatus.labelRes(): Int = when (this) {
+internal fun ShiftStatus.labelRes(): Int = when (this) {
     ShiftStatus.PLANNED -> R.string.shifts_status_planned
     ShiftStatus.SWAP_PENDING -> R.string.shifts_status_swap_pending
     ShiftStatus.CONFIRMED -> R.string.shifts_status_confirmed
 }
 
-private fun ShiftAssignment.effectiveDateMillis(overrides: List<DeliveryCalendarOverride>): Long =
+internal fun ShiftAssignment.effectiveDateMillis(overrides: List<DeliveryCalendarOverride>): Long =
     if (type == ShiftType.DELIVERY) {
         overrides.firstOrNull { it.weekKey == dateMillis.toWeekKey() }?.deliveryDateMillis ?: dateMillis
     } else {
         dateMillis
     }
 
-private fun ShiftAssignment.toSummaryLine(
+fun ShiftAssignment.toSummaryLine(
     members: List<Member>,
     overrides: List<DeliveryCalendarOverride>,
 ): String = "${effectiveDateMillis(overrides).toLocalizedDateTime()} · ${assignedUserIds.toMemberNames(members)}"
 
 @Composable
-private fun ShiftAssignment.leftBoardLines(overrides: List<DeliveryCalendarOverride>): List<ShiftBoardLine> = when (type) {
+internal fun ShiftAssignment.leftBoardLines(overrides: List<DeliveryCalendarOverride>): List<ShiftBoardLine> = when (type) {
     ShiftType.DELIVERY -> {
         val localDate = effectiveDateMillis(overrides).toLocalDate()
         listOf(
@@ -4590,7 +2031,7 @@ private fun ShiftAssignment.leftBoardLines(overrides: List<DeliveryCalendarOverr
     }
 }
 
-private fun ShiftAssignment.primaryBoardNames(members: List<Member>): List<String> = when (type) {
+internal fun ShiftAssignment.primaryBoardNames(members: List<Member>): List<String> = when (type) {
     ShiftType.DELIVERY -> buildList {
         assignedUserIds.firstOrNull()?.let { add(members.displayNameFor(it)) }
         add(helperUserId?.let { members.displayNameFor(it) } ?: "—")
@@ -4606,7 +2047,7 @@ private fun ShiftAssignment.primaryBoardNames(members: List<Member>): List<Strin
         }
 }
 
-private fun ShiftAssignment.canBeRequestedBy(
+internal fun ShiftAssignment.canBeRequestedBy(
     currentMemberId: String,
     overrides: List<DeliveryCalendarOverride>,
 ): Boolean = when (type) {
@@ -4616,7 +2057,7 @@ private fun ShiftAssignment.canBeRequestedBy(
         assignedUserIds.contains(currentMemberId)
 }
 
-private fun ShiftAssignment.highlightedBoardNameIndex(currentMemberId: String): Int? = when (type) {
+internal fun ShiftAssignment.highlightedBoardNameIndex(currentMemberId: String): Int? = when (type) {
     ShiftType.DELIVERY -> when {
         assignedUserIds.firstOrNull() == currentMemberId -> 0
         helperUserId == currentMemberId -> 1
@@ -4637,13 +2078,13 @@ private fun ShiftSwapRequest.availableResponses(): List<com.reguerta.user.domain
     responses.filter { it.status == com.reguerta.user.domain.shifts.ShiftSwapResponseStatus.AVAILABLE }
 
 @StringRes
-private fun ShiftSwapRequestStatus.labelRes(): Int = when (this) {
+internal fun ShiftSwapRequestStatus.labelRes(): Int = when (this) {
     ShiftSwapRequestStatus.OPEN -> R.string.shift_swap_request_status_open
     ShiftSwapRequestStatus.CANCELLED -> R.string.shift_swap_request_status_cancelled
     ShiftSwapRequestStatus.APPLIED -> R.string.shift_swap_request_status_applied
 }
 
-private fun ShiftBoardSegment.toShiftType(): ShiftType = when (this) {
+internal fun ShiftBoardSegment.toShiftType(): ShiftType = when (this) {
     ShiftBoardSegment.DELIVERY -> ShiftType.DELIVERY
     ShiftBoardSegment.MARKET -> ShiftType.MARKET
 }
@@ -4653,7 +2094,7 @@ private fun Long.toLocalDate(): LocalDate =
         .atZone(ZoneId.systemDefault())
         .toLocalDate()
 
-private fun Long.toWeekKey(): String {
+internal fun Long.toWeekKey(): String {
     val localDate = toLocalDate()
     val weekFields = WeekFields.ISO
     val week = localDate.get(weekFields.weekOfWeekBasedYear())
@@ -5069,78 +2510,6 @@ private fun AdminToolsCard(
 }
 
 @Composable
-private fun OperationalModules(
-    modulesEnabled: Boolean,
-    canOpenProducts: Boolean,
-    myOrderFreshnessState: MyOrderFreshnessUiState,
-    onRetryMyOrderFreshness: () -> Unit,
-    onOpenProducts: () -> Unit,
-    onOpenShifts: () -> Unit,
-    disabledMessage: String? = null,
-) {
-    Card {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(10.dp),
-        ) {
-            Text(stringResource(R.string.operational_modules_title))
-            Button(
-                onClick = {},
-                enabled = modulesEnabled && myOrderFreshnessState == MyOrderFreshnessUiState.Ready,
-                modifier = Modifier.fillMaxWidth(),
-            ) {
-                Text(stringResource(R.string.module_my_order))
-            }
-            Button(onClick = onOpenProducts, enabled = modulesEnabled && canOpenProducts, modifier = Modifier.fillMaxWidth()) {
-                Text(stringResource(R.string.module_catalog))
-            }
-            Button(onClick = onOpenShifts, enabled = modulesEnabled, modifier = Modifier.fillMaxWidth()) {
-                Text(stringResource(R.string.module_shifts))
-            }
-
-            if (!modulesEnabled && disabledMessage != null) {
-                Text(
-                    text = disabledMessage,
-                    style = MaterialTheme.typography.bodySmall,
-                )
-            }
-
-            when (myOrderFreshnessState) {
-                MyOrderFreshnessUiState.Checking -> {
-                    Text(
-                        text = stringResource(R.string.my_order_freshness_checking),
-                        style = MaterialTheme.typography.bodySmall,
-                    )
-                }
-
-                MyOrderFreshnessUiState.TimedOut,
-                MyOrderFreshnessUiState.Unavailable,
-                    -> {
-                        Text(
-                            text = stringResource(R.string.my_order_freshness_error_title),
-                            style = MaterialTheme.typography.titleSmall,
-                            fontWeight = FontWeight.SemiBold,
-                        )
-                        Text(
-                            text = stringResource(R.string.my_order_freshness_error_message),
-                            style = MaterialTheme.typography.bodySmall,
-                        )
-                        TextButton(onClick = onRetryMyOrderFreshness) {
-                            Text(stringResource(R.string.my_order_freshness_retry))
-                        }
-                    }
-
-                MyOrderFreshnessUiState.Idle,
-                MyOrderFreshnessUiState.Ready,
-                    -> Unit
-            }
-        }
-    }
-}
-
-@Composable
 private fun RoleCheckboxRow(
     checked: Boolean,
     label: String,
@@ -5247,7 +2616,7 @@ private fun isValidEmail(email: String): Boolean =
 private fun isValidPassword(password: String): Boolean =
     password.length in PasswordMinLength..PasswordMaxLength
 
-private fun NotificationAudience.labelRes(): Int =
+fun NotificationAudience.labelRes(): Int =
     when (this) {
         NotificationAudience.ALL -> R.string.notifications_target_all
         NotificationAudience.MEMBERS -> R.string.notifications_target_members
@@ -5255,7 +2624,7 @@ private fun NotificationAudience.labelRes(): Int =
         NotificationAudience.ADMINS -> R.string.notifications_target_admins
     }
 
-private fun NotificationEvent.audienceLabelRes(): Int =
+fun NotificationEvent.audienceLabelRes(): Int =
     when {
         target == "all" -> R.string.notifications_target_all
         target == "users" -> R.string.notifications_target_users
@@ -5265,43 +2634,16 @@ private fun NotificationEvent.audienceLabelRes(): Int =
         else -> R.string.notifications_target_all
     }
 
-private fun Long.toLocalizedDateTime(): String =
+fun Long.toLocalizedDateTime(): String =
     DateFormat.getDateTimeInstance(DateFormat.MEDIUM, DateFormat.SHORT).format(java.util.Date(this))
 
-private fun Long.toLocalizedDateOnly(): String =
+internal fun Long.toLocalizedDateOnly(): String =
     java.text.SimpleDateFormat("d MMM yyyy", Locale.forLanguageTag("es-ES")).format(java.util.Date(this))
 
-private fun ShiftAssignment.toShiftSwapDisplayLabel(
+internal fun ShiftAssignment.toShiftSwapDisplayLabel(
     memberId: String?,
     overrides: List<DeliveryCalendarOverride>,
 ): String = effectiveDateMillis(overrides).toLocalizedDateOnly()
-
-private fun Long.toDeliveryWeekday(): DeliveryWeekday =
-    when (toLocalDate().dayOfWeek) {
-        java.time.DayOfWeek.MONDAY -> DeliveryWeekday.MONDAY
-        java.time.DayOfWeek.TUESDAY -> DeliveryWeekday.TUESDAY
-        java.time.DayOfWeek.WEDNESDAY -> DeliveryWeekday.WEDNESDAY
-        java.time.DayOfWeek.THURSDAY -> DeliveryWeekday.THURSDAY
-        java.time.DayOfWeek.FRIDAY -> DeliveryWeekday.FRIDAY
-        java.time.DayOfWeek.SATURDAY -> DeliveryWeekday.SATURDAY
-        java.time.DayOfWeek.SUNDAY -> DeliveryWeekday.SUNDAY
-    }
-
-private fun DeliveryWeekday.toSpanishLabel(): String = when (this) {
-    DeliveryWeekday.MONDAY -> "Lunes"
-    DeliveryWeekday.TUESDAY -> "Martes"
-    DeliveryWeekday.WEDNESDAY -> "Miercoles"
-    DeliveryWeekday.THURSDAY -> "Jueves"
-    DeliveryWeekday.FRIDAY -> "Viernes"
-    DeliveryWeekday.SATURDAY -> "Sabado"
-    DeliveryWeekday.SUNDAY -> "Domingo"
-}
-
-private fun DeliveryWeekday.previous(): DeliveryWeekday =
-    DeliveryWeekday.entries[(ordinal + DeliveryWeekday.entries.size - 1) % DeliveryWeekday.entries.size]
-
-private fun DeliveryWeekday.next(): DeliveryWeekday =
-    DeliveryWeekday.entries[(ordinal + 1) % DeliveryWeekday.entries.size]
 
 private sealed interface StartupGateUiState {
     data object Checking : StartupGateUiState
